all.sh: merge "full config" and "SSLv3" tests
SSLv3 is enabled in the full config, so we don't need a separate build just
for it. Also, for compat.sh and all.sh run only the tests the make sense (no
need to re-test all mordern TLS 1.2 ciphersuites in compat.sh), but make sure
to test all of them (previously RC4/simple-DES/NULL ciphersuites were not
tested with SSLv3).

Fix a test name in ssl-opt.sh so that filtering on SSLv3 works.