all.sh: make CFLAGS=... doesn't work with cmake
It's not the way you're supposed to use CFLAGS with cmake, see
https://github.com/ARMmbed/mbedtls/pull/1241

Remove -std=c99 -pedantic as it doesn't work for some modules in the library
(eg net_sockets), and there is a (correct) test for the parts of the library
that are supposed to build with it later anyway.

While at it, change to use make, not only because it's clearer how CFLAGS
work, but also because that way we test the default config with gcc+cmake and
the full one with clang+make, covering both build systems with those two basic
builds.

Also, 'make test' only runs the test, it doesn't build them ('make tests'
does), so CFLAGS are irrelevant there.