Split stream ciphers out of ssl_decrypt_buf()
This is the first step in splitting mode-specific part of that function into 4
different functions, in order to make code flow more readable (currently 3
flows are interleaved in a non-trivial way in a nearly 500 lines function).

During that work, some code will be temporarily duplicated. It will later be
extracted to helper functions in order to remove duplication.