Add stack.sh
Changes:

- NO_PTR_CALL avoids pointer calls (f_rng, grp->modp, platform_zeroize)
for the benefit of wcs.py

- there was funny things going on with mbedtls_mpi_shift_l_mod where
gcc would notice that MOD_ADD in common with mbedtls_mpi_add_mod and
generate a sub-function mbedtls_mpi_shift_l_mod.part.0 that would be an
alias to mbedtls_mpi_add_mod.part.0 (that is, both at the same address
in the object file), but in the dfinish file would only show the tree
for mbedtls_mpi_add_mod.part.0, and since wcs.py doesn't know about the
aliasing, it was complaining about not knowing the call tree of
mbedtls_mpi_shift_l_mod.part.0... The proper fix would of course to make
wcs.py recognize this kind of aliasing, but since this is a one-off,
just change the definition of the function so that GCC no longer has a
subfunction to alias. (GCC also noticed that the last argument was
always 1, so suppress it.)

Note: values in libc.msu are quite arbitrary, and probably optimistic
for some of them.